getMessages
