+++ packages/next-intl/src/server/staticConfig.tsx
@@ -1,5 +1,5 @@
 // @ts-expect-error Should we provide a default? Probably.
 import staticConfig from 'next-intl/config';
-import NextI18nConfig from './NextI18nConfig';
+import NextIntlConfig from './NextIntlConfig';
 
-export default staticConfig as NextI18nConfig;
+export default staticConfig as NextIntlConfig;
